Change --output-proto-directory behavior (#75)
`--output-proto-directory` behavior seems to be not very useful when working with protofetch modules. It changes the default output directory, so it only works if the protofetch.toml file doesn't specify `proto_out_dir`. A more useful behavior seems to be to override the output directory, whether it's set in protofetch.toml or not.

This is technically a breaking change, but it is only breaking if someone has `proto_out_dir` set and is passing a different value to `--output-proto-directory` (which is then ignored).