Prevent concurrent cache access (#135)
`libgit2` fails if we try to access the same cached repository
concurrently. This is however very likely to happen when multiple
projects in cargo workspace use protofetch to fetch from the same
source.

This PR makes protofetch acquire a lock on the cache directory before
accessing it, so only one instance can work with the cache at a time.
The lock is not granular, but it should be good enough.